requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
cryptography==42.0.5
Deprecated==1.2.14
easy-flask-swagger==1.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
flask-swagger-ui==4.11.1
frozenlist==1.4.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.5
packaging==23.2
protobuf==4.21.12
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
twilio==9.0.2
typing_extensions==4.8.0
urllib3==2.0.7
vonage==3.13.0
vonage-jwt==1.1.0
Werkzeug==3.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
